Refactored code and removed old task-add module
