+++ src/client/flogo/flow/flow-routing.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { EmptyDetailComponent } from './empty-detail/empty-detail.component';
-import { FlogoFlowsDetailTasksComponent } from './task-add';
 import { FlogoFlowsDetailTasksDetailComponent } from './task-detail';
 import { FlowComponent } from './flow.component';
 import { FlowDataResolver } from './flow-data.resolver';
@@ -16,7 +15,6 @@ const flowRoutes: Routes = [
     },
     children: [
       { path: '', component: EmptyDetailComponent },
-      { path: 'task/add', component: FlogoFlowsDetailTasksComponent },
       { path: 'task/:id', component: FlogoFlowsDetailTasksDetailComponent }
     ]
   },
