--- src/client/flogo/flow/task-add-new/task-add.component.html
@@ -1,32 +1,0 @@
-<div class="main">
-  <h2 class="title">{{"TASKS:ADD-ACTIVITY" | translate}}</h2>
-  <div class="search">
-    <input class="flogo-input-search search-input" type="text"
-           placeholder="{{ 'ADD:FILTER-TASKS' | translate }}"
-           (keyup)="filterActivities($event.target.value)">
-  </div>
-  <div *ngIf="filteredActivities$ | async as activities" class="scrollable-content">
-    <ul class="activity-list" *ngIf="activities.length > 0; else emptyCase">
-      <li class="activity-list-item qa-activities" *ngFor="let activity of activities" (click)="selectActivity(activity.ref)">
-        {{activity.title}}
-      </li>
-    </ul>
-    <ng-template #emptyCase>
-      <p class="empty-text qa-no-tasks">{{ 'ADD:NO-TASKS' | translate }}</p>
-    </ng-template>
-  </div>
-</div>
-<div class="install">
-  <div class="install-activity qa-install-activity" (click)="handleInstallerWindow(true)">
-    {{ 'FLOWS-DETAILS-TASKS:INSTALL-NEW-ACTIVITY' | translate }}
-  </div>
-</div>
-<flogo-flow-installer [profileType]="(options.appAndFlowInfo$ | async).appProfileType"
-                      [installType]="'activity'"
-                      [isActivated]="isInstallOpen"
-                      (isActivatedChange)="handleInstallerWindow($event)"
-                      (installed)="afterActivityInstalled($event)"></flogo-flow-installer>
-<flogo-flow-sub-flow *ngIf="isSubflowOpen"
-                     [appId]="(options.appAndFlowInfo$ | async).appId"
-                     [currentFlow]="(options.appAndFlowInfo$ | async).actionId"
-                     (flowSelected)="handleFlowSelection($event)"></flogo-flow-sub-flow>
