--- src/client/flogo/flow/task-add-new/task-add.component.less
@@ -1,60 +1,0 @@
-@import (reference) '../../../assets/uxpl/dropdown';
-@import (reference) '../../../assets/_typography';
-@import '../../../assets/_variables';
-
-:host {
-  .uxpl-dropdown-container();
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  width: 250px;
-  height: 285px;
-}
-
-.main {
-  padding: 10px 10px 0;
-  border-bottom: 1px solid @color-medium-gray;
-}
-
-.title {
-  .secondary-header();
-  margin: 0 0 10px;
-}
-
-.search {
-  text-align: center;
-}
-
-.search-input {
-  width: 100%;
-}
-
-.scrollable-content {
-  height: 170px;
-  overflow-y: auto;
-}
-
-.empty-text {
-  margin: 1em;
-}
-
-.activity-list {
-  .uxpl-dropdown();
-}
-
-.activity-list-item {
-  .uxpl-dropdown__item();
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.install {
-  background: @lighter-gray;
-}
-
-.install-activity {
-  .uxpl-dropdown__item();
-  padding-left: 26px;
-  padding-right: 26px;
-}
