+++ src/client/flogo/flow/task-add/add-activity.directive.ts
