+++ src/client/flogo/flow/task-add/core/filter-activities-by.ts
@@ -1,7 +1,7 @@
 import {Observable, ReplaySubject} from 'rxjs';
 import {map, shareReplay, withLatestFrom} from 'rxjs/operators';
 import {isEmpty, filter} from 'lodash';
-import {Activity} from '../task-add.component';
+import {Activity} from './task-add-options';
 
 const filterList = (activities, filterText) => {
   if (filterText && !isEmpty(filterText)) {
