+++ src/client/flogo/flow/task-add/index.ts
@@ -1,2 +1,3 @@
 export {TaskAddModule} from './task-add.module';
-export {FlogoFlowsDetailTasksComponent} from './task-add.component';
+export {TaskAddComponent} from './task-add.component';
+export * from './core/task-add-options';
