--- src/client/flogo/flow/task-add/install/index.ts
@@ -1,1 +1,0 @@
-export {InstallModule} from './install.module';
