--- src/client/flogo/flow/task-add/install/install.component.html
@@ -1,13 +1,0 @@
-<div>
-  <button type="button" class="flogo-common-edit-panel__btn-download"
-          (click)="openModal()">
-    <i class="flogo-icon-download"></i>
-    {{ 'FLOWS-DETAILS-TASKS:INSTALL-NEW-ACTIVITY' | translate }}
-  </button>
-
-  <flogo-flow-installer
-    [profileType]="profileType"
-    [installType]="'activity'" [(isActivated)]="isActivated"
-                   (installed)="onInstalledAction($event)"></flogo-flow-installer>
-
-</div>
