--- src/client/flogo/flow/task-add/install/install.component.less
@@ -1,5 +1,0 @@
-:host {
-  padding: 0 19px 21px 19px;
-}
-
-/* Modal styles need to be global scoped, located in main css */
