+++ src/client/flogo/flow/task-add/task-add.component.html
@@ -1,33 +1,32 @@
-<div class="flogo-common-edit-panel flogo-flows-detail-activities">
-
-  <div class="flogo-common-edit-panel__head">
-    <div class="flogo-common-edit-panel__head-wrapper">
-      <i class="flogo-icon-add"></i>
-      <h3 class="flogo-common-edit-panel__head-title"> {{ 'TASKS:ADD-ACTIVITY' | translate }} </h3>
-      <div class="flogo-common-edit-panel__head-subtitle">
-        {{'TASKS:HINT-ADD-ACTIVITY' | translate}}
-      </div>
-    </div>
-    <flogo-flow-task-add-install
-      [profileType]="profileType"
-      (flogoOnInstalled)="onInstalledAction($event)"></flogo-flow-task-add-install>
+<div class="main">
+  <h2 class="title">{{"TASKS:ADD-ACTIVITY" | translate}}</h2>
+  <div class="search">
+    <input class="flogo-input-search search-input" type="text"
+           placeholder="{{ 'ADD:FILTER-TASKS' | translate }}"
+           (keyup)="filterActivities($event.target.value)">
   </div>
-
-  <div class="flogo-common-edit-panel__container">
-    <div class="filter-box">
-      <input class="flogo-input-search filter-box__input" type="text" [(ngModel)]="filterQuery"
-             placeholder="{{ 'ADD:FILTER-TASKS' | translate }}">
-    </div>
-
-
-    <ul class="flogo-common-edit-panel__tiles-list">
-      <li class="flogo-common-edit-panel__tiles-list-tile" *ngFor="let task of filteredTasks"
-          (click)="manageAddTaskMsg(task)">{{task.name}}
+  <div *ngIf="filteredActivities$ | async as activities" class="scrollable-content">
+    <ul class="activity-list" *ngIf="activities.length > 0; else emptyCase">
+      <li class="activity-list-item qa-activities" *ngFor="let activity of activities" (click)="selectActivity(activity.ref)">
+        {{activity.title}}
       </li>
     </ul>
+    <ng-template #emptyCase>
+      <p class="empty-text qa-no-tasks">{{ 'ADD:NO-TASKS' | translate }}</p>
+    </ng-template>
   </div>
-  <flogo-flow-sub-flow *ngIf="showFlowsList"
-                       [appId]="appId"
-                       [currentFlow]="currentFlowId"
-                       (flowSelected)="handleFlowSelection($event)"></flogo-flow-sub-flow>
 </div>
+<div class="install">
+  <div class="install-activity qa-install-activity" (click)="handleInstallerWindow(true)">
+    {{ 'FLOWS-DETAILS-TASKS:INSTALL-NEW-ACTIVITY' | translate }}
+  </div>
+</div>
+<flogo-flow-installer [profileType]="(options.appAndFlowInfo$ | async).appProfileType"
+                      [installType]="'activity'"
+                      [isActivated]="isInstallOpen"
+                      (isActivatedChange)="handleInstallerWindow($event)"
+                      (installed)="afterActivityInstalled($event)"></flogo-flow-installer>
+<flogo-flow-sub-flow *ngIf="isSubflowOpen"
+                     [appId]="(options.appAndFlowInfo$ | async).appId"
+                     [currentFlow]="(options.appAndFlowInfo$ | async).actionId"
+                     (flowSelected)="handleFlowSelection($event)"></flogo-flow-sub-flow>
